requirements.txt
flask==2.3.2
python-dotenv==1.0.0
openai>=1.51.0,<2.0.0
PyGithub==2.1.1
pydantic==2.4.2
astroid==3.0.1
ell-ai==0.0.13
